layers: Move class BASE_NODE to its own file
