+++ BUILD.gn
@@ -90,6 +90,7 @@ core_validation_sources = [
   "layers/core_validation.h",
   "layers/core_validation_error_enums.h",
   "layers/core_validation_types.h",
+  "layers/base_node.h",
   "layers/core_error_location.cpp",
   "layers/core_error_location.h",
   "layers/descriptor_sets.cpp",
