+++ layers/CMakeLists.txt
@@ -156,6 +156,7 @@ set(CORE_VALIDATION_LIBRARY_FILES
     core_error_location.h
     core_error_location.cpp
     core_validation_types.h
+    base_node.h
     drawdispatch.cpp
     convert_to_renderpass2.cpp
     descriptor_sets.cpp
